network/usage: allow ability to specify if network's ipaddress usage need to be hidden
This adds a new API arg only accessible to admins to specify if they
want the network's ip address usage hidden. This then saves this setting
in the network_details table for a network, and the listUsageRecords
API response creator checks for an IP address if it needs to be exported
or skipped/hidden.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>