Do not add comments with double quotes
The rules added in iptables to whitelist traffic would be destroyed and
recreated every time an operation in the VPC would trigger a iptables
reconfiguration. This would reset the network counters and break the
functionnality.

This PR aims at fixing this issue, by correctly having the VR keep the
whitelist rules if they are already applied in iptables.

Original Author: @ccontini
Originally added in !133