Fix slow vm creation when large sf snapshot count
Slow VM creation times are experienced when using solid fire managed
storage. It is suspected that this is caused by calls to getUsedBytes in
the solid fire storage driver. Storage used by snapshots is not being
computed efficiently resulting in a large number of DB queries.

getUsedBytes is called a number of times during the vm creation process.
This PR proposes to refactor it so that a constant number of queries are
used to compute it. We do one query each to the volumes table, the
snapshot_store_ref table and the template_spool_ref table instead of
querying the respective *_details tables by object id to get the size of
each type of object.

Steps to reproduce:
Attempt to create a VM in a Cloudstack instance that has a large number
of snapshots
Expected behaviour:
VM creation should happen in a less than a few minutes

Actual behaviour:
VM creation takes around 5 minutes even with small images

Original Author: @skattoju4
Original PR: https://github.com/apache/cloudstack/pull/3282